pyproject.toml
[project]
name = "jobhunt"
version = "0.1.0"
description = "AI-powered job search aggregator for the German & European market"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0,<22.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "alembic>=1.14,<2.0",
    "aiosqlite>=0.20,<1.0",
    "asyncpg>=0.30,<1.0",
    "python-jobspy>=1.1,<2.0",
    "anthropic>=0.42,<1.0",
    "aiohttp>=3.11,<4.0",
    "apscheduler>=3.10,<4.0",
    "pydantic-settings>=2.7,<3.0",
    "python-dotenv>=1.0,<2.0",
    "python-dateutil>=2.9,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 120
target-version = "py311"
